+++ src/Schema/Quoter.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Schema;
 
+use Yiisoft\Db\Exception\InvalidArgumentException;
+use Yiisoft\Db\Expression\ExpressionInterface;
 use function addcslashes;
 use function explode;
 use function implode;
@@ -32,6 +34,45 @@ class Quoter implements QuoterInterface
         private array|string $tableQuoteCharacter,
         private string $tablePrefix = ''
     ) {
+    }
+
+    public function cleanUpTableNames(array $tableNames): array
+    {
+        $cleanedUpTableNames = [];
+        $pattern = <<<PATTERN
+        ~^\s*((?:['"`\[]|{{).*?(?:['"`\]]|}})|\(.*?\)|.*?)(?:(?:\s+(?:as\s+)?)((?:['"`\[]|{{).*?(?:['"`\]]|}})|.*?))?\s*$~iux
+        PATTERN;
+
+        /** @psalm-var array<array-key, ExpressionInterface|string> $tableNames */
+        foreach ($tableNames as $alias => $tableName) {
+            if (is_string($tableName) && !is_string($alias)) {
+                if (preg_match($pattern, $tableName, $matches)) {
+                    if (isset($matches[2])) {
+                        [, $tableName, $alias] = $matches;
+                    } else {
+                        $tableName = $alias = $matches[1];
+                    }
+                }
+            }
+
+            if (!is_string($alias)) {
+                throw new InvalidArgumentException(
+                    'To use Expression in from() method, pass it in array format with alias.'
+                );
+            }
+
+            if (is_string($tableName)) {
+                $cleanedUpTableNames[$this->ensureNameQuoted($alias)] = $this->ensureNameQuoted($tableName);
+            } elseif ($tableName instanceof ExpressionInterface) {
+                $cleanedUpTableNames[$this->ensureNameQuoted($alias)] = $tableName;
+            } else {
+                throw new InvalidArgumentException(
+                    'Use ExpressionInterface without cast to string as object of tableName'
+                );
+            }
+        }
+
+        return $cleanedUpTableNames;
     }
 
     public function getTableNameParts(string $name, bool $withColumn = false): array
